rename: type 및 api 이름 변경
